Refactor compute_costing function to separate module
