+++ cads_processing_api_service/costing.py
@@ -14,11 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License
 
+from typing import Any
+
+import cads_adaptors
 import cads_catalogue
 import fastapi
 import ogc_api_processes_fastapi.models
 
-from . import adaptors, db_utils, models, utils
+from . import adaptors, costing, db_utils, models, utils
 
 
 def estimate_costs(
@@ -34,5 +37,24 @@ def estimate_costs(
             resource_id=process_id, table=table, session=catalogue_session
         )
     adaptor_properties = adaptors.get_adaptor_properties(dataset)
-    costing = adaptors.compute_costing(request, adaptor_properties)
-    return costing
+    costing_info = costing.compute_costing(request, adaptor_properties)
+    return costing_info
+
+
+def compute_costing(
+    request: dict[str, Any],
+    adaptor_properties: dict[str, Any],
+) -> models.Costing:
+    adaptor: cads_adaptors.AbstractAdaptor = adaptors.instantiate_adaptor(
+        adaptor_properties=adaptor_properties
+    )
+    costs: dict[str, float] = adaptor.estimate_costs(request=request)
+    costing_config: dict[str, Any] = adaptor_properties["config"].get("costing", {})
+    max_costs: dict[str, Any] = costing_config.get("max_costs", {})
+    max_costs_exceeded = {}
+    for max_cost_id, max_cost_value in max_costs.items():
+        if max_cost_id in costs.keys():
+            if costs[max_cost_id] > max_cost_value:
+                max_costs_exceeded[max_cost_id] = max_cost_value
+    costing_info = models.Costing(costs=costs, max_costs_exceeded=max_costs_exceeded)
+    return costing_info
